Clean up source impact report: move renderer, fix ident access, drop JSON round-trip
Move _render_remote_section from the CLI module into reports/source_impact.py,
where all other rendering helpers live. Add remote_summary as an optional
parameter to render_markdown and render_json so callers pass it in directly
instead of parsing a freshly-serialized string to inject a key.

Also remove the duplicate print(infile) in anatomy.compute_cliques_for_impact_report
and fix ident.get("i") → ident["i"] in _final_compendium_breakdown, consistent
with every other compendium reader in the codebase.
